Fix formatting and add VaultToken

+++ contracts/interfaces/ITarotCallee.sol
@@ -1,6 +1,16 @@
 pragma solidity >=0.5.0;
 
 interface ITarotCallee {
-    function tarotBorrow(address sender, address borrower, uint borrowAmount, bytes calldata data) external;
-    function tarotRedeem(address sender, uint redeemAmount, bytes calldata data) external;
+    function tarotBorrow(
+        address sender,
+        address borrower,
+        uint256 borrowAmount,
+        bytes calldata data
+    ) external;
+
+    function tarotRedeem(
+        address sender,
+        uint256 redeemAmount,
+        bytes calldata data
+    ) external;
 }
