+++ test/Contracts/tarot-core/interfaces/IBorrowTracker.sol
@@ -1,5 +1,9 @@
 pragma solidity >=0.5.0;
 
 interface IBorrowTracker {
-	function trackBorrow(address borrower, uint borrowBalance, uint borrowIndex) external;
+    function trackBorrow(
+        address borrower,
+        uint256 borrowBalance,
+        uint256 borrowIndex
+    ) external;
 }
