+++ test/Contracts/uniswap-v2-core/test/ERC20.sol
@@ -1,9 +1,9 @@
 pragma solidity =0.5.16;
 
-import '../UniswapV2ERC20.sol';
+import "../UniswapV2ERC20.sol";
 
 contract ERC20 is UniswapV2ERC20 {
-    constructor(uint _totalSupply) public {
+    constructor(uint256 _totalSupply) public {
         _mint(msg.sender, _totalSupply);
     }
 }
